+++ trello-clone/src/main/resources/db/migration/V9__add_cardlists_table.sql
@@ -2,7 +2,7 @@ create table cardlists
 (
     id               uuid primary key,
     board_id         uuid                                      not null references boards (id),
-    position         smallint                                  not null,
+    position         integer                                   not null,
 
     unique (board_id, position),
